+++ src/Engine/AppContext.hpp
@@ -4,6 +4,7 @@
 #include "Common.hpp"
 #include "IO/MouseInputState.hpp"
 #include "IO/MouseCaptureMode.hpp"
+#include "terrain_platform.h"
 
 namespace Terrain { namespace Engine {
     class EXPORT AppContext
@@ -12,8 +13,8 @@ namespace Terrain { namespace Engine {
         // input
         virtual void updateInputState() = 0;
         virtual const IO::MouseInputState &getMouseState(int inputControllerId) const = 0;
-        virtual const unsigned short &getPressedMouseButtons(int inputControllerId) const = 0;
-        virtual const unsigned long long &getPressedKeys(int inputControllerId) const = 0;
+        virtual const uint8 &getPressedMouseButtons(int inputControllerId) const = 0;
+        virtual const uint64 &getPressedKeys(int inputControllerId) const = 0;
         virtual void setMouseCaptureMode(IO::MouseCaptureMode mode) = 0;
     };
 }}
