+++ src/Engine/IO/InputManager.hpp
@@ -22,8 +22,8 @@ namespace Terrain { namespace Engine { namespace IO {
         {
             int count;
             std::vector<MouseInputState> mouse;
-            std::vector<unsigned short> pressedMouseButtons;
-            std::vector<unsigned long long> pressedKeys;
+            std::vector<uint8> pressedMouseButtons;
+            std::vector<uint64> pressedKeys;
 
             InputState() : count(0)
             {
@@ -40,32 +40,30 @@ namespace Terrain { namespace Engine { namespace IO {
 
         bool isKeyDown(int inputControllerId, Key key) const
         {
-            unsigned long long pressedKeys = ctx.getPressedKeys(inputControllerId);
-            return (pressedKeys & static_cast<unsigned long long>(key)) != 0;
+            uint64 pressedKeys = ctx.getPressedKeys(inputControllerId);
+            return (pressedKeys & static_cast<uint64>(key)) != 0;
         }
 
         bool isNewKeyPress(int inputControllerId, Key key) const
         {
-            unsigned long long currentPressedKeys = ctx.getPressedKeys(inputControllerId);
-            unsigned long long prevPressedKeys = prevInputState.pressedKeys[inputControllerId];
-            return ((currentPressedKeys & static_cast<unsigned long long>(key)) != 0)
-                && ((prevPressedKeys & static_cast<unsigned long long>(key)) == 0);
+            uint64 currentPressedKeys = ctx.getPressedKeys(inputControllerId);
+            uint64 prevPressedKeys = prevInputState.pressedKeys[inputControllerId];
+            return ((currentPressedKeys & static_cast<uint64>(key)) != 0)
+                && ((prevPressedKeys & static_cast<uint64>(key)) == 0);
         }
 
         bool isMouseButtonDown(int inputControllerId, MouseButton button) const
         {
-            unsigned short pressedButtons = ctx.getPressedMouseButtons(inputControllerId);
-            return (pressedButtons & static_cast<unsigned short>(button)) != 0;
+            uint8 pressedButtons = ctx.getPressedMouseButtons(inputControllerId);
+            return (pressedButtons & static_cast<uint8>(button)) != 0;
         }
 
         bool isNewMouseButtonPress(int inputControllerId, MouseButton button) const
         {
-            unsigned short currentPressedButtons =
-                ctx.getPressedMouseButtons(inputControllerId);
-            unsigned short prevPressedButtons =
-                prevInputState.pressedMouseButtons[inputControllerId];
-            return ((currentPressedButtons & static_cast<unsigned long long>(button)) != 0)
-                && ((prevPressedButtons & static_cast<unsigned long long>(button)) == 0);
+            uint8 currentPressedButtons = ctx.getPressedMouseButtons(inputControllerId);
+            uint8 prevPressedButtons = prevInputState.pressedMouseButtons[inputControllerId];
+            return ((currentPressedButtons & static_cast<uint8>(button)) != 0)
+                && ((prevPressedButtons & static_cast<uint8>(button)) == 0);
         }
 
         void addInputController();
