+++ src/Engine/IO/MouseButton.hpp
@@ -2,9 +2,10 @@
 #define IO_MOUSEBUTTON_HPP
 
 #include "../Common.hpp"
+#include "../terrain_platform.h"
 
 namespace Terrain { namespace Engine { namespace IO {
-    enum class MouseButton : unsigned short
+    enum class MouseButton : uint8
     {
         Left = 1 << 0,
         Middle = 1 << 1,
